test: cover support summary cases without dependencies
